Move and go-fmt tokeniser.
